requirements.txt
pandas==2.2.2
streamlit==1.35.0
openpyxl==3.1.2
Pillow==9.0.0